requirements.txt
beautifulsoup4==4.11.0
Pillow==10.0.0
selenium==3.141.0
streamlit==1.21.0
seleniumbase
webdriver-manager